# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/revisions/{revisionId}:submit": {
      "post": {
        "operationId": "submitRevision",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRevisionDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {
                        "revisionId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "layerId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "snapshotId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "generation": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/v1/admin/revisions/{revisionId}:approve": {
      "post": {
        "operationId": "approveRevision",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {
                        "revisionId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "layerId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "snapshotId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "generation": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/v1/admin/revisions/{revisionId}:request-changes": {
      "post": {
        "operationId": "requestRevisionChanges",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestChangesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {
                        "revisionId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "layerId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "snapshotId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "generation": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/v1/admin/revisions/{revisionId}:publish": {
      "post": {
        "operationId": "publishRevision",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRevisionDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {
                        "revisionId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "layerId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "snapshotId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "generation": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/v1/admin/layers/{layerId}:rollback": {
      "post": {
        "operationId": "rollbackLayer",
        "parameters": [
          {
            "name": "layerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {
                        "revisionId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "layerId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "snapshotId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "generation": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "workflow"
        ]
      }
    },
...
  },
  "info": {
    "title": "DanangMap API",
    "description": "DanangMap v2 spatial CMS contract",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-danangmap_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
...
      "SubmitRevisionDto": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string",
            "maxLength": 4000
          },
          "reviewerNote": {
            "type": "string",
            "maxLength": 4000
          }
        },
        "required": [
          "summary"
        ]
      },
      "WorkflowCommentDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "maxLength": 4000
          }
        }
      },
      "RequestChangesDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "maxLength": 4000
          }
        },
        "required": [
          "comment"
        ]
      },
      "PublishRevisionDto": {
        "type": "object",
        "properties": {
          "releaseNote": {
            "type": "string",
            "maxLength": 4000
          }
        },
        "required": [
          "releaseNote"
        ]
      },
      "RollbackDto": {
        "type": "object",
        "properties": {
          "targetSnapshotId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "maxLength": 4000
          }
        },
        "required": [
          "targetSnapshotId",
          "reason"
        ]
      }
    }
  }
}