# src/workflow/workflow.controller.ts
import {
  Body,
  Controller,
  Headers,
  HttpCode,
  Param,
  ParseUUIDPipe,
  Post,
  Req,
  UseGuards,
} from '@nestjs/common';
import { ApiCookieAuth, ApiOperation, ApiTags } from '@nestjs/swagger';
import type { RequestWithContext } from '../common/http/request-context';
import { apiJsonResponse, workflowResultSchema } from '../common/openapi/response-schemas';
import { Principal, Roles } from '../identity/auth.decorators';
import { CsrfGuard, RolesGuard, SessionGuard } from '../identity/auth.guards';
import { requireIdempotencyKey } from '../layers/etag';
import {
  PublishRevisionDto,
  RequestChangesDto,
  RollbackDto,
  SubmitRevisionDto,
  WorkflowCommentDto,
} from '../layers/layer.dto';
import { WorkflowService } from './workflow.service';

@ApiTags('workflow')
@ApiCookieAuth('adminSession')
@Controller({ path: 'admin', version: '1' })
@UseGuards(SessionGuard, RolesGuard, CsrfGuard)
export class WorkflowController {
  constructor(private readonly workflow: WorkflowService) {}

  @Post('revisions/:revisionId\\:submit')
  @HttpCode(202)
  @Roles('editor')
  @ApiOperation({ operationId: 'submitRevision' })
  @apiJsonResponse(202, workflowResultSchema)
  submit(
    @Param('revisionId', ParseUUIDPipe) revisionId: string,
    @Body() dto: SubmitRevisionDto,
    @Headers('idempotency-key') key: string | undefined,
    @Req() request: RequestWithContext,
    @Principal() principal: NonNullable<RequestWithContext['principal']>,
  ) {
    requireIdempotencyKey(key);
    return this.workflow.submit(revisionId, dto, principal, request.requestId);
  }

  @Post('revisions/:revisionId\\:approve')
  @Roles('reviewer')
  @ApiOperation({ operationId: 'approveRevision' })
  @apiJsonResponse(201, workflowResultSchema)
  approve(
    @Param('revisionId', ParseUUIDPipe) revisionId: string,
    @Body() dto: WorkflowCommentDto,
    @Headers('idempotency-key') key: string | undefined,
    @Req() request: RequestWithContext,
    @Principal() principal: NonNullable<RequestWithContext['principal']>,
  ) {
    requireIdempotencyKey(key);
    return this.workflow.approve(revisionId, dto, principal, request.requestId);
  }

  @Post('revisions/:revisionId\\:request-changes')
  @Roles('reviewer')
  @ApiOperation({ operationId: 'requestRevisionChanges' })
  @apiJsonResponse(201, workflowResultSchema)
  requestChanges(
    @Param('revisionId', ParseUUIDPipe) revisionId: string,
    @Body() dto: RequestChangesDto,
    @Headers('idempotency-key') key: string | undefined,
    @Req() request: RequestWithContext,
    @Principal() principal: NonNullable<RequestWithContext['principal']>,
  ) {
    requireIdempotencyKey(key);
    return this.workflow.requestChanges(revisionId, dto, principal, request.requestId);
  }

  @Post('revisions/:revisionId\\:publish')
  @HttpCode(202)
  @Roles('publisher')
  @ApiOperation({ operationId: 'publishRevision' })
  @apiJsonResponse(202, workflowResultSchema)
  publish(
    @Param('revisionId', ParseUUIDPipe) revisionId: string,
    @Body() dto: PublishRevisionDto,
    @Headers('idempotency-key') key: string | undefined,
    @Req() request: RequestWithContext,
    @Principal() principal: NonNullable<RequestWithContext['principal']>,
  ) {
    requireIdempotencyKey(key);
    return this.workflow.publish(revisionId, dto, principal, request.requestId);
  }

  @Post('layers/:layerId\\:rollback')
  @Roles('publisher')
  @ApiOperation({ operationId: 'rollbackLayer' })
  @apiJsonResponse(201, workflowResultSchema)
  rollback(
    @Param('layerId', ParseUUIDPipe) layerId: string,
    @Body() dto: RollbackDto,
    @Headers('idempotency-key') key: string | undefined,
    @Req() request: RequestWithContext,
    @Principal() principal: NonNullable<RequestWithContext['principal']>,
  ) {
    requireIdempotencyKey(key);
    return this.workflow.rollback(layerId, dto, principal, request.requestId);
  }
}